fscrypt: Adding additional documentation
This commit moves most of the documentation about contributing to
fscrypt into CONTRIBUTING.md and updates the legal disclaimer.

It also updates the README.md to include all of fscrypt's planned
functionality and dependencies. Finally, the makefile is updated to
include more documentation, versioning support, and a different location
for the output file.

Change-Id: Ib7be98d41bc06dd12b02e42addf06e12a940235a