vendor: add in external packages for use with CLI
crypto/ssh/terminal gets us information about the terminal, such as
its dimensions and state. It also allows us to manipulate the terminal.
For example, we can put the terminal in raw mode when a passphrase needs
to be entered.

This commit also add the github.com/urfave/cli package which we will use
to write the command line tool component of fscrypt. This tool allows
for us to easily use commands and subcommands.

Note that this is actually the upstream repository with two PRs applied:

https://github.com/urfave/cli/pull/629
https://github.com/urfave/cli/pull/630

They fix bugs in the handling of custom usage errors.

Change-Id: I2d3ba967b1bce8f73440e3a06df4eaba7ab96c19