README: remove note about planned commands
These would still be nice to add.  However, the mention of them in the
README is misleading because people reading it might come away with the
impression that there is currently no way to back up fscrypt metadata or
to recover directories -- which isn't true.  (The fscrypt metadata is
just a directory which can be backed up like any other directory.  And
'fscrypt encrypt' already offers to generate a recovery passphrase when
the directory and protector are on different filesystems.)

Just remove this note; it doesn't really add any value.